// src/dataverse-actions-functions-power-automate-flows-samples/src/generated/models/WorkflowsModel.ts
export const Workflowsbusinessprocesstype = {
  0: 'BusinessFlow',
  1: 'TaskFlow'
} as const;
export type Workflowsbusinessprocesstype = keyof typeof Workflowsbusinessprocesstype;
export const Workflowscategory = {
  0: 'Workflow',
  1: 'Dialog',
  2: 'BusinessRule',
  3: 'Action',
  4: 'BusinessProcessFlow',
  5: 'ModernFlow',
  6: 'DesktopFlow',
  7: 'AIFlow',
  9000: 'WebClientAPIFlow'
} as const;
export type Workflowscategory = keyof typeof Workflowscategory;
export const Workflowscomponentstate = {
  0: 'Published',
  1: 'Unpublished',
  2: 'Deleted',
  3: 'DeletedUnpublished'
} as const;
export type Workflowscomponentstate = keyof typeof Workflowscomponentstate;
export const Workflowscreatestage = {
  20: 'Pre_operation',
  40: 'Post_operation'
} as const;
export type Workflowscreatestage = keyof typeof Workflowscreatestage;
export const Workflowsdeletestage = {
  20: 'Pre_operation',
  40: 'Post_operation'
} as const;
export type Workflowsdeletestage = keyof typeof Workflowsdeletestage;
export const Workflowsmode = {
  0: 'Background',
  1: 'Real_time'
} as const;
export type Workflowsmode = keyof typeof Workflowsmode;
export const Workflowsmodernflowtype = {
  0: 'PowerAutomateFlow',
  1: 'CopilotStudioFlow',
  2: 'M365CopilotAgentFlow'
} as const;
export type Workflowsmodernflowtype = keyof typeof Workflowsmodernflowtype;
export const Workflowsprocesstriggerscope = {
  1: 'Form',
  2: 'Entity'
} as const;
export type Workflowsprocesstriggerscope = keyof typeof Workflowsprocesstriggerscope;
export const Workflowsrunas = {
  0: 'Owner',
  1: 'CallingUser'
} as const;
export type Workflowsrunas = keyof typeof Workflowsrunas;
export const Workflowsscope = {
  1: 'User',
  2: 'BusinessUnit',
  3: 'Parent_ChildBusinessUnits',
  4: 'Organization'
} as const;
export type Workflowsscope = keyof typeof Workflowsscope;
export const Workflowsstatecode = {
  0: 'Draft',
  1: 'Activated',
  2: 'Suspended'
} as const;
export type Workflowsstatecode = keyof typeof Workflowsstatecode;
export const Workflowsstatuscode = {
  1: 'Draft',
  2: 'Activated',
  3: 'CompanyDLPViolation'
} as const;
export type Workflowsstatuscode = keyof typeof Workflowsstatuscode;
export const Workflowsthrottlingbehavior = {
  0: 'None',
  1: 'TenantPool',
  2: 'CopilotStudio'
} as const;
export type Workflowsthrottlingbehavior = keyof typeof Workflowsthrottlingbehavior;
export const Workflowstype = {
  1: 'Definition',
  2: 'Activation',
  3: 'Template'
} as const;
export type Workflowstype = keyof typeof Workflowstype;
export const Workflowsuiflowtype = {
  0: 'Windowsrecorder_V1_',
  1: 'SeleniumIDE',
  2: 'PowerAutomateDesktop',
  3: 'Test',
  101: 'Recording'
} as const;
export type Workflowsuiflowtype = keyof typeof Workflowsuiflowtype;
export const Workflowsupdatestage = {
  20: 'Pre_operation',
  40: 'Post_operation'
} as const;
export type Workflowsupdatestage = keyof typeof Workflowsupdatestage;

export interface WorkflowsBase {
  asyncautodelete: boolean;
  billingcontext?: string;
  businessprocesstype?: Workflowsbusinessprocesstype;
  category: Workflowscategory;
  claims?: string;
  clientdata?: string;
  connectionreferences?: string;
  createmetadata?: string;
  createstage?: Workflowscreatestage;
  credentials?: string;
  definition?: string;
  deletestage?: Workflowsdeletestage;
  dependencies?: string;
  description?: string;
  desktopflowmodules?: string;
  dynamicssolutioncontext?: string;
  entityimage?: string;
  formid?: string;
  inputparameters?: string;
  inputs?: string;
  introducedversion?: string;
  iscustomizable: string;
  iscustomprocessingstepallowedforotherpublishers: string;
  istransacted?: boolean;
  languagecode?: number;
  "Licensee@odata.bind"?: string;
  "LicenseEntitledBy@odata.bind"?: string;
  metadata?: string;
  mode: Workflowsmode;
  modernflowtype?: Workflowsmodernflowtype;
  modifymetadata?: string;
  name: string;
  ondemand?: boolean;
  outputs?: string;
  ownerid: string;
  owneridtype: string;
  planverified?: boolean;
  primaryentity: string;
  processorder?: number;
  processroleassignment?: string;
  processtriggerformid?: string;
  processtriggerscope?: Workflowsprocesstriggerscope;
  rank?: number;
  rendererobjecttypecode?: string;
  resourcecontainer?: string;
  resourceid?: string;
  runas?: Workflowsrunas;
  schemaversion?: string;
  scope: Workflowsscope;
  statecode: Workflowsstatecode;
  statuscode?: Workflowsstatuscode;
  subprocess?: boolean;
  suspensionreasondetails?: string;
  syncworkflowlogonfailure: boolean;
  throttlingbehavior?: Workflowsthrottlingbehavior;
  triggeroncreate?: boolean;
  triggerondelete?: boolean;
  triggeronupdateattributelist?: string;
  type: Workflowstype;
  uiflowtype?: Workflowsuiflowtype;
  uniquename?: string;
  updatestage?: Workflowsupdatestage;
  workflowid: string;
  xaml: string;
}

export interface Workflows extends WorkflowsBase {
  activeworkflowidname?: string;
  asyncautodeletename?: string;
  businessprocesstypename?: string;
  categoryname?: string;
  clientdataiscompressed?: boolean;
  clientdataiscompressedname?: string;
  componentstate: Workflowscomponentstate;
  createdbyname?: string;
  createdbyyominame?: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame?: string;
  createstagename?: string;
  deletestagename?: string;
  entityimage_timestamp?: number;
  entityimage_url?: string;
  entityimageid?: string;
  iscrmuiworkflow?: boolean;
  ismanaged: boolean;
  ismanagedname?: string;
  istransactedname?: string;
  licenseename?: string;
  licenseentitledbyname?: string;
  licenseeyominame?: string;
  modename?: string;
  modernflowtypename?: string;
  modifiedbyname?: string;
  modifiedbyyominame?: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame?: string;
  ondemandname?: string;
  overwritetime: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname?: string;
  parentworkflowidname?: string;
  planverifiedname?: string;
  primaryentityname?: string;
  processtriggerscopename?: string;
  runasname?: string;
  scopename?: string;
  solutionid: string;
  statecodename?: string;
  statuscodename?: string;
  subprocessname?: string;
  supportingsolutionid?: string;
  syncworkflowlogonfailurename?: string;
  throttlingbehaviorname?: string;
  triggeroncreatename?: string;
  triggerondeletename?: string;
  trustedaccess?: boolean;
  trustedaccessname?: string;
  typename?: string;
  uidata?: string;
  uiflowtypename?: string;
  updatestagename?: string;
  versionnumber?: number;
  workflowidunique: string;
  activeworkflowid?: object;
  _activeworkflowid_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  licensee?: object;
  _licensee_value?: string;
  licenseentitledby?: object;
  _licenseentitledby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  parentworkflowid?: object;
  _parentworkflowid_value?: string;
  plugintypeid?: object;
  _plugintypeid_value?: string;
  sdkmessageid?: object;
  _sdkmessageid_value?: string;
}